prefabの時だけcopyしていたが、シーンの時もコピーするように変えた。 #423
* 不要になったRecordDisposerを使わない。
* CopyVRMComponents を BoneNormalizer 内部に移動。
